products references userModel and moved models to folders
