+++ modules/orderService.js
@@ -1,6 +1,5 @@
 var mongoose = require("mongoose");
-var bcrypt = require("bcryptjs");
-var ObjectId = require("mongodb").ObjectId;
+const Orders = require("./Models/OrderModel");
 
 async function doConnect() {
 	await mongoose.connect("mongodb://localhost/eez", {
@@ -18,16 +17,6 @@ function parseResponse(response) {
 	return parsed;
 }
 
-const Orders = mongoose.model(
-	"Orders",
-	new mongoose.Schema({
-		SellerID: { type: String },
-		destAddress: { type: String },
-		CC: { type: Number },
-		status: { type: String },
-		total: { type: String }
-	})
-);
 
 module.exports.getAllOrders = sID => {
 	return new Promise((resolve, reject) => {
